feat: extract gateway_me_mcp and make register_authenticated_tools a backward-compatible no-op
Move the get_me/get_my_notes/get_my_posts MCP tools out of gateway_oss
into a new gateway_me_mcp capability package registered via the
substack_gateway.capabilities entry-point group. register_authenticated_tools
in gateway_oss/mcp/app.py is now an empty no-op, kept only so a downstream
extension that still calls it does not double-register the me tools and
crash FastMCP on duplicate tool names.
