+++ packages/gateway_me_mcp/pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "gateway_me_mcp"
+version = "3.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_mcp_common", "gateway_profiles", "gateway_notes", "gateway_posts", "fastmcp>=3.4.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+me_mcp = "gateway_me_mcp:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_me_mcp"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_mcp_common = { workspace = true }
+gateway_profiles = { workspace = true }
+gateway_notes = { workspace = true }
+gateway_posts = { workspace = true }
