+++ packages/gateway_oss/pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
   "gateway_profiles_rest",
   "gateway_profiles_mcp",
   "gateway_me_rest",
+  "gateway_me_mcp",
   "gateway_rest_common",
   "gateway_mcp_common",
 ]
@@ -59,5 +60,6 @@ gateway_comments_mcp = { workspace = true }
 gateway_profiles_rest = { workspace = true }
 gateway_profiles_mcp = { workspace = true }
 gateway_me_rest = { workspace = true }
+gateway_me_mcp = { workspace = true }
 gateway_rest_common = { workspace = true }
 gateway_mcp_common = { workspace = true }
